aula-3-autentication-mobile-and-web

+++ server/src/routes/memories.ts
@@ -3,11 +3,18 @@ import { z } from "zod";
 import { prisma } from '../lib/prisma'
 
 export async function memoriesRoutes(app: FastifyInstance) {
-  app.get("/memories", async () => {
+  app.addHook('preHandler', async (request) => {
+    await request.jwtVerify();
+  })
+
+  app.get("/memories", async (request) => {
     const memories = await prisma.memory.findMany({
         orderBy: {
             createdAt: 'asc',
         },
+        where: {
+            userId: request.user.sub,
+        }
     });
     
     return memories.map((memory) => ({
@@ -17,7 +24,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
     }))
   });
 
-  app.get("/memories/:id", async (request) => {
+  app.get("/memories/:id", async (request, reply) => {
     const paramsSchema = z.object({
         id: z.string().uuid(),
     })
@@ -30,8 +37,12 @@ export async function memoriesRoutes(app: FastifyInstance) {
         }
     })
 
-    return memory;
-  });
+    if(!memory.isPublic && memory.userId !== request.user.sub) {
+        reply.status(401).send()
+    }
+
+        return memory;
+    });
 
   app.post("/memories", async (request) => {
     const bodySchema = z.object({
@@ -47,14 +58,14 @@ export async function memoriesRoutes(app: FastifyInstance) {
             content,
             isPublic,
             coverUrl,
-            userId: "1"
+            userId: request.user.sub,
         }
     })
 
     return memory;
   });
 
-  app.put("/memories/:id", async (request) => {
+  app.put("/memories/:id", async (request, reply) => {
     const paramsSchema = z.object({
         id: z.string().uuid(),
     })
@@ -69,7 +80,17 @@ export async function memoriesRoutes(app: FastifyInstance) {
 
     const { content, isPublic, coverUrl } = bodySchema.parse(request.body);
 
-    const memory = await prisma.memory.update({
+    let memory = await prisma.memory.findUnique({
+        where: {
+            id,
+        }
+    })
+
+    if(memory?.userId !== request.user.sub) {
+        return reply.status(401).send()
+    }
+
+    memory = await prisma.memory.update({
         where: {
             id,
         },
@@ -83,14 +104,24 @@ export async function memoriesRoutes(app: FastifyInstance) {
     return memory;
   });
 
-  app.delete("/memories/:id", async (request) => {
+  app.delete("/memories/:id", async (request, reply) => {
     const paramsSchema = z.object({
         id: z.string().uuid(),
     })
 
     const { id } = paramsSchema.parse(request.params);
 
-    const memory = await prisma.memory.delete({
+    const memory = await prisma.memory.findUniqueOrThrow({
+        where: {
+            id,
+        },
+    })
+
+    if(memory.userId !== request.user.sub) {
+        return reply.status(401).send()
+    }
+
+     await prisma.memory.delete({
         where: {
             id,
         }
